# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Pharmacy & Doctor Finder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Wilaya {
  id: number;
  code: string;
  name: string;
  name_ar: string;
  region: string;
  pharmacy_count?: number;
  doctor_count?: number;
}

export interface Pharmacy {
  id: number;
  name: string;
  name_ar?: string;
  address: string;
  commune?: string;
  wilaya_id: number;
  wilaya_name: string;
  phone?: string;
  phone2?: string;
  email?: string;
  latitude?: number;
  longitude?: number;
  hours_open?: string;
  hours_close?: string;
  is_24h?: boolean;
  on_duty: boolean;
  on_duty_until?: string;
  rating?: number;
  review_count?: number;
  services?: string[];
  image_url?: string;
}

export interface PharmacyListResponse {
  data: Pharmacy[];
  total: number;
  page: number;
  limit: number;
  pages: number;
}

export interface Doctor {
  id: number;
  first_name: string;
  last_name: string;
  title?: string;
  specialty: string;
  specialty_ar?: string;
  sub_specialty?: string;
  address: string;
  commune?: string;
  wilaya_id: number;
  wilaya_name: string;
  phone?: string;
  phone2?: string;
  email?: string;
  latitude?: number;
  longitude?: number;
  consultation_hours?: string;
  consultation_fee?: number;
  accepts_cnas?: boolean;
  accepts_casnos?: boolean;
  languages?: string[];
  rating?: number;
  review_count?: number;
  experience_years?: number;
  bio?: string;
  image_url?: string;
  available_today?: boolean;
}

export interface DoctorListResponse {
  data: Doctor[];
  total: number;
  page: number;
  limit: number;
  pages: number;
}

export type ListPharmaciesParams = {
  wilaya_id?: number;
  on_duty?: boolean;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListDoctorsParams = {
  wilaya_id?: number;
  specialty?: string;
  search?: string;
  page?: number;
  limit?: number;
};