# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Pharmacy & Doctor Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: number;
  first_name: string;
  last_name: string;
  title?: string;
  specialty: string;
  specialty_ar?: string;
  sub_specialty?: string;
  address: string;
  commune?: string;
  wilaya_id: number;
  wilaya_name: string;
  phone?: string;
  phone2?: string;
  email?: string;
  latitude?: number;
  longitude?: number;
  consultation_hours?: string;
  consultation_fee?: number;
  accepts_cnas?: boolean;
  accepts_casnos?: boolean;
  languages?: string[];
  rating?: number;
  review_count?: number;
  experience_years?: number;
  bio?: string;
  image_url?: string;
  available_today?: boolean;
}