# lib/api-zod/src/generated/types/pharmacy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Pharmacy & Doctor Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface Pharmacy {
  id: number;
  name: string;
  name_ar?: string;
  address: string;
  commune?: string;
  wilaya_id: number;
  wilaya_name: string;
  phone?: string;
  phone2?: string;
  email?: string;
  latitude?: number;
  longitude?: number;
  hours_open?: string;
  hours_close?: string;
  is_24h?: boolean;
  on_duty: boolean;
  on_duty_until?: string;
  rating?: number;
  review_count?: number;
  services?: string[];
  image_url?: string;
}