# lib/api-zod/src/generated/types/pharmacyListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Pharmacy & Doctor Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { Pharmacy } from "./pharmacy";

export interface PharmacyListResponse {
  data: Pharmacy[];
  total: number;
  page: number;
  limit: number;
  pages: number;
}