# lib/api-zod/src/generated/types/wilaya.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Pharmacy & Doctor Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface Wilaya {
  id: number;
  code: string;
  name: string;
  name_ar: string;
  region: string;
  pharmacy_count?: number;
  doctor_count?: number;
}